requirements.txt
anyio==4.12.1
certifi==2026.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
python-telegram-bot==20.7
sniffio==1.3.1
typing_extensions==4.15.0
